pkg/storageos: separate into multiple files
